% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_example.R
\name{hsr_example}
\alias{hsr_example}
\title{Get path to an example Hospital-Specific Report (HSR)}
\usage{
hsr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL, the example file will be listed.}
}
\description{
Provides the location of the example HSR that comes with the package for users to access with package functions. This file is for FY 2025 program year downloaded on 11/8/2025 from https://qualitynet.cms.gov/inpatient/hrrp/reports.
}
\examples{
hsr_example()
hsr_example("999999_HRRP_Readmissions_HSR.xlsx")
}
