% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_extract_cohort_summary.R
\name{hsr_extract_cohort_summary}
\alias{hsr_extract_cohort_summary}
\title{Extract cohort summary information from a Hospital-Specific Report (HSR)}
\usage{
hsr_extract_cohort_summary(file)
}
\arguments{
\item{file}{File path to a report.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the full Table 2 parsed from the report.
}
\description{
Parses the Table 2 cohort summary from the HSR, including (but not limited to) the discharge/readmission volumes,
predicted/expected readmission rates, peer group medians, and DRG ratios.
}
\examples{
my_report <- hsr_mock_reports("FY2021_HRRP_MockHSR.xlsx")
cohort_summary <- hsr_extract_cohort_summary(my_report)
cohort_summary
}
