% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_payment_summary.R
\name{hsr_extract_payment_summary}
\alias{hsr_extract_payment_summary}
\alias{hsr_count_dual_eligible_stays}
\alias{hsr_count_total_stays}
\alias{hsr_dual_proportion}
\alias{hsr_peer_group}
\alias{hsr_neutrality_modifier}
\alias{hsr_payment_adjustment_factor}
\alias{hsr_payment_penalty}
\title{Extract payment summary information a Hospital-Specific Report (HSR)}
\usage{
hsr_extract_payment_summary(file)

hsr_count_dual_eligible_stays(file)

hsr_count_total_stays(file)

hsr_dual_proportion(file)

hsr_peer_group(file)

hsr_neutrality_modifier(file)

hsr_payment_adjustment_factor(file)

hsr_payment_penalty(file)
}
\arguments{
\item{file}{File path to a report}
}
\value{
A \code{\link[tibble]{tibble}} containing the columns \code{Measure} and \code{Value} for the metric and value, respectively.
}
\description{
Parses Table 1 of the Hospital-Specific Report, either altogether in a single output or each component individually.
}
\examples{
my_report <- hsr_mock_reports("FY2021_HRRP_MockHSR.xlsx")
hsr_extract_payment_summary(my_report)
}
