% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_read.R
\name{is_hsr}
\alias{is_hsr}
\title{Check whether an object is a parsed HSR}
\usage{
is_hsr(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \verb{\\"readmit_hsr\\"}, otherwise
\code{FALSE}.
}
\description{
Check whether an object is a parsed HSR
}
\examples{
report <- hsr_mock_reports("FY2025_HRRP_MockHSR.xlsx")
parsed <- hsr_read(report)
is_hsr(parsed)
is_hsr(report)
}
