% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_parse.R
\name{is_hsr_bundle}
\alias{is_hsr_bundle}
\title{Check whether an object is a parsed HSR bundle}
\usage{
is_hsr_bundle(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \verb{\\"readmit_hsr_bundle\\"},
otherwise \code{FALSE}.
}
\description{
Check whether an object is a parsed HSR bundle
}
