% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_source.R
\name{is_hsr_source}
\alias{is_hsr_source}
\title{Check whether an object is an HSR source bundle}
\usage{
is_hsr_source(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \verb{\\"readmit_hsr_source\\"},
otherwise \code{FALSE}.
}
\description{
Check whether an object is an HSR source bundle
}
