% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc_datasets.R
\name{pdc_datasets}
\alias{pdc_datasets}
\title{List the datasets available in the Provider Data Catalog (PDC)}
\usage{
pdc_datasets(topics = NULL)
}
\arguments{
\item{topics}{A topic to list datasets for (see \code{\link{pdc_topics}}). If NULL, all datasets will be returned.}
}
\value{
A \code{\link[tibble]{tibble}} containing dataset metadata. Specifically, an entry from the \code{datasetid} column can be used in \code{\link{pdc_read}}.
}
\description{
Provides the names, identifiers, and other associated metadata of datasets available from the CMS Provider Data Catalog (PDC) (https://data.cms.gov/provider-data/). This can be used to explore dataset contents/metadata and/or be used for identifying datasets to import in \code{\link{pdc_read}}.
}
\examples{
pdc_topics()
pdc_datasets("Hospitals")
}
