% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc_read.R
\name{pdc_read}
\alias{pdc_read}
\title{Import a dataset from the Provider Data Catalog (PDC)}
\usage{
pdc_read(datasetid, ...)
}
\arguments{
\item{datasetid}{A dataset identifier (see \code{\link{pdc_datasets}}).}

\item{...}{Additional arguments passed to \code{\link[readr]{read_csv}}}
}
\value{
A \code{\link[tibble]{tibble}} containing the requested dataset.
}
\description{
Imports a complete dataset from the CMS Provider Data Catalog (PDC) (https://data.cms.gov/provider-data/). Dataset identifiers can be discovered on the webpage and/or with \code{\link{pdc_datasets}}.
}
\examples{
pdc_datasets("Hospitals") |> dplyr::filter(stringr::str_detect(title, "(?i)readmission"))
pdc_read("9n3s-kdb3")
}
