% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc_topics.R
\name{pdc_topics}
\alias{pdc_topics}
\title{List the topics from the Provider Data Catalog (PDC)}
\usage{
pdc_topics()
}
\value{
A character vector
}
\description{
Provides the collection of topics from the CMS Provider Data Catalog (PDC) (https://data.cms.gov/provider-data/) in which various datasets are available for. This is typically used to retrieve a topic selection for downstream use in \code{\link{pdc_datasets}}.
}
\examples{
pdc_topics()
}
