% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmit-package.R
\docType{package}
\name{readmit-package}
\alias{readmit}
\alias{readmit-package}
\title{readmit: R Interface to the Readmission Risk Pool}
\description{
\code{readmit} provides an R interface to the Readmission Risk Pool platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centralstatz.github.io/readmit/}
  \item \url{https://github.com/centralstatz/readmit}
  \item \url{https://github.com/centralstatz/readmission-risk-pool-platform}
  \item Report bugs at \url{https://github.com/centralstatz/readmit/issues}
}

}
\author{
\strong{Maintainer}: Alex Zajichek \email{alex@centralstatz.com}

Other contributors:
\itemize{
  \item CentralStatz Statistical & Data Sciences LLC [copyright holder]
}

}
\keyword{internal}
