% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_extract_coefficients.R
\name{hsr_extract_coefficients}
\alias{hsr_extract_coefficients}
\title{Extract model coefficients from a Hospital-Specific Report (HSR)}
\usage{
hsr_extract_coefficients(file, cohort)
}
\arguments{
\item{file}{File path to a report}

\item{cohort}{Cohort to extract the coefficients for. One of \code{c("AMI", "COPD", "HF", "PN", "CABG", "HK")}}
}
\value{
A \code{\link[tibble]{tibble}} containing the columns \code{Factor} and \code{Value} for the model term and coefficient value, respectively (on the linear-predictor scale).
}
\description{
Parses out the regression coefficients from the logistic regression model used by CMS to estimate discharge-level readmission risk, including the hospital-level and hospital average intercept terms.
}
\examples{
\dontrun{
my_report <- "Readmissions_HSR.xlsx"
hsr_extract_coefficients(my_report, "AMI")
}
}
