% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsr_extract_coefficients.R
\name{hsr_extract_coefficients}
\alias{hsr_extract_coefficients}
\title{Extract model coefficients from a Hospital-Specific Report (HSR)}
\usage{
hsr_extract_coefficients(file, cohort)
}
\arguments{
\item{file}{File path to a report}

\item{cohort}{Program cohort to extract the coefficients for. One of \code{c("AMI", "COPD", "HF", "PN", "CABG", "HK")}}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
Extract model coefficients from a Hospital-Specific Report (HSR)
}
\examples{
\dontrun{
my_report <- "Readmissions_HSR.xlsx"
hsr_extract_coefficients(my_report, "AMI")
}
}
