% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmit-package.R
\docType{package}
\name{readmit-package}
\alias{readmit}
\alias{readmit-package}
\title{readmit: Tools for navigating, managing, and analyzing hospital readmissions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains toolkits and datasets to make managing hospital readmissions easier and more transparent. Particular context is given to the Hospital Readmissions Reduction Program (HRRP) by the Center for Medicare & Medicaid Services (CMS) for understanding program details and analyzing reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centralstatz.github.io/readmit/}
}

}
\author{
\strong{Maintainer}: Alex Zajichek \email{alex@centralstatz.com}

}
\keyword{internal}
