% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmit-package.R
\docType{package}
\name{readmit-package}
\alias{readmit}
\alias{readmit-package}
\title{readmit: Tools for Working with Hospital Readmissions Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains tools for working with and analyzing hospital readmissions data. The package provides utilities for components of the Hospital Readmissions Reduction Program (HRRP), including program timeline functions, Hospital-Specific Report (HSR) helpers, and general importing tools for the Provider Data Catalog (PDC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centralstatz.github.io/readmit/}
  \item \url{https://github.com/centralstatz/readmit}
  \item Report bugs at \url{https://github.com/centralstatz/readmit/issues}
}

}
\author{
\strong{Maintainer}: Alex Zajichek \email{alex@centralstatz.com} [copyright holder]

}
\keyword{internal}
